se agregaron cambios esteticos a check in cart payment planes se migro la base de datos y se agregaron nuevas funcionalidades ademas de agregar la api de google maps

+++ app/src/main/java/cl/gymtastic/app/data/local/entity/CartItemEntity.kt
@@ -8,5 +8,5 @@ data class CartItemEntity(
     @PrimaryKey(autoGenerate = true) val id: Long = 0,
     val productId: Long,
     val qty: Int,
-    val unitPrice: Double
+    val unitPrice: Int
 )
